+++ src/lib/pages/validator-details/components/bonded-token-changes/VotingPowerChart.tsx
@@ -1,30 +1,47 @@
 import { Box, Flex, Heading, Text } from "@chakra-ui/react";
+import type { BigSource } from "big.js";
 import type { ScriptableContext, TooltipModel } from "chart.js";
 
+import { useCelatoneApp } from "lib/app-provider";
 import { LineChart } from "lib/components/chart/LineChart";
+import { Loading } from "lib/components/Loading";
+import { ErrorFetching } from "lib/components/state";
+import { useAssetInfos } from "lib/services/assetService";
+import { useValidatorHistoricalPowers } from "lib/services/validatorService";
+import type { Token, U, ValidatorAddr } from "lib/types";
+import {
+  formatHHmm,
+  formatUTokenWithPrecision,
+  getTokenLabel,
+} from "lib/utils";
 
 interface VotingPowerChartProps {
-  denom: string;
+  validatorAddress: ValidatorAddr;
 }
 
-export const VotingPowerChart = ({ denom }: VotingPowerChartProps) => {
-  const labels = [
-    "00:00",
-    "01:00",
-    "02:00",
-    "03:00",
-    "04:00",
-    "05:00",
-    "06:00",
-    "07:00",
-    "08:00",
-    "09:00",
-    "10:00",
-    "11:00",
-    "12:00",
-  ];
+export const VotingPowerChart = ({
+  validatorAddress,
+}: VotingPowerChartProps) => {
+  const {
+    chainConfig: {
+      extra: { singleStakingDenom },
+    },
+  } = useCelatoneApp();
+  const { data: assetInfos, isLoading: isAssetInfosLoading } = useAssetInfos({
+    withPrices: false,
+  });
 
-  const dateLabels = labels.map((label) => {
+  const { data: historicalPowers, isLoading } =
+    useValidatorHistoricalPowers(validatorAddress);
+
+  if (isLoading || isAssetInfosLoading) return <Loading />;
+  if (!historicalPowers) return <ErrorFetching dataName="historical powers" />;
+
+  const labels = historicalPowers?.items.map((item) =>
+    formatHHmm(item.hourRoundedTimestamp as Date)
+  );
+
+  const dateLabels = labels?.map((label) => {
     const [hours, minutes] = label.split(":");
     const date = new Date();
     date.setHours(parseInt(hours, 10));
@@ -35,10 +52,8 @@ export const VotingPowerChart = ({ denom }: VotingPowerChartProps) => {
   });
 
   const dataset = {
-    data: [
-      2000, 2500, 1800, 2100, 1200, 1500, 1800, 2000, 2500, 1800, 2100, 1200,
-      1500,
-    ],
+    data:
+      historicalPowers?.items.map((item) => item.votingPower.toNumber()) ?? [],
     borderColor: "#D8BEFC",
     backgroundColor: (context: ScriptableContext<"line">) => {
       const { ctx } = context.chart;
@@ -54,7 +69,29 @@ export const VotingPowerChart = ({ denom }: VotingPowerChartProps) => {
     pointHoverBorderColor: "#D8BEFC",
   };
 
-  const currentPrice = dataset.data[dataset.data.length - 1].toFixed(1);
+  const assetInfo = singleStakingDenom
+    ? assetInfos?.[singleStakingDenom]
+    : undefined;
+
+  const handleFormatValue = (value: string | number, isSuffix = true) =>
+    formatUTokenWithPrecision(
+      value as U<Token<BigSource>>,
+      assetInfo?.precision ?? 0,
+      isSuffix,
+      2
+    );
+
+  const isDatasetContainsData = dataset && dataset.data.length > 0;
+
+  const currency = singleStakingDenom
+    ? `${getTokenLabel(singleStakingDenom, assetInfo?.symbol)}`
+    : "";
+  const currentPrice = isDatasetContainsData
+    ? handleFormatValue(dataset.data[dataset.data.length - 1])
+    : "";
+  const diffInLast24Hr = isDatasetContainsData
+    ? dataset.data[dataset.data.length - 1] - dataset.data[0]
+    : 0;
 
   const customizeTooltip = (tooltip: TooltipModel<"line">) => {
     const { raw, dataIndex } = tooltip.dataPoints[0];
@@ -62,8 +99,10 @@ export const VotingPowerChart = ({ denom }: VotingPowerChartProps) => {
     return `
       <div style="padding: 8px 12px;">
         <div style="font-weight: 700;">
-          <h1 style="font-size: 12px; color: #ADADC2;">Bonded Tokens</h1>
-          <p style="font-size: 16px; color: #F7F2FE; white-space: nowrap;">${Number(raw).toFixed(1)} ${denom}</p>
+          <h1 style="font-size: 12px; color: #ADADC2;">${
+            singleStakingDenom ? "Bonded Token" : "Voting Powers"
+          }</h1>
+          <p style="font-size: 16px; color: #F7F2FE; white-space: nowrap;">${handleFormatValue(raw as number, false)} ${currency}</p>
         </div>
         <hr style="margin-top: 8px; color: #68688A;"/>
         <p style="margin-top: 8px; font-size: 12px; color: #F7F2FE; white-space: nowrap;">${dateLabels[dataIndex]}</p>
@@ -84,15 +123,25 @@ export const VotingPowerChart = ({ denom }: VotingPowerChartProps) => {
       w="100%"
     >
       <Flex gap={2} direction="column" w={250} minW={250}>
-        <Heading variant="h6">Current Bonded Token</Heading>
+        <Heading variant="h6">
+          {singleStakingDenom
+            ? "Current Bonded Token"
+            : "Current Voting Powers"}
+        </Heading>
         <Heading variant="h5" fontWeight={600}>
-          {currentPrice} {denom}
+          {currentPrice} {currency}
         </Heading>
         <Text variant="body1">
-          <Text as="span" fontWeight={700} color="success.main">
-            +24.02
+          <Text
+            as="span"
+            fontWeight={700}
+            color={diffInLast24Hr >= 0 ? "success.main" : "error.main"}
+          >
+            {diffInLast24Hr >= 0
+              ? `+${handleFormatValue(diffInLast24Hr)}`
+              : `-${handleFormatValue(-diffInLast24Hr)}`}
           </Text>{" "}
-          {denom} in last 24 hr
+          {currency} in last 24 hr
         </Text>
       </Flex>
       <Box w="100%" h="272px" id="voting-power-chart-container">
@@ -100,6 +149,7 @@ export const VotingPowerChart = ({ denom }: VotingPowerChartProps) => {
           labels={labels}
           dataset={dataset}
           customizeTooltip={customizeTooltip}
+          customizeYAxisTicks={handleFormatValue}
         />
       </Box>
     </Flex>